Moved `format_hex_address` to `utils.py`
